processors.py
import asyncio
from PyPDF2 import PdfReader
import docx
import requests
from bs4 import BeautifulSoup
import speech_recognition as sr
from pydub import AudioSegment
from io import BytesIO
from utils import extract_text_from_image, summarize_text

async def process_document(doc):
    file = await doc.get_file()
    file_content = await file.download_as_bytearray()
    
    if doc.file_name.endswith('.pdf'):
        return await process_pdf(BytesIO(file_content))
    elif doc.file_name.endswith('.docx'):
        return await process_docx(BytesIO(file_content))
    else:
        return "Unsupported document type"

async def process_pdf(file_obj):
    reader = PdfReader(file_obj)
    if len(reader.pages) > 30:
        return "PDF is too long (over 30 pages)"
    
    text = ""
    for page in reader.pages:
        text += page.extract_text()
    
    if not text:
        # If no text extracted, it might be a scanned PDF
        return await process_scanned_pdf(file_obj)
    
    summary = await summarize_text(text)
    return summary

async def process_scanned_pdf(file_obj):
    # Use OCR or GPT-4 Vision here
    text = await extract_text_from_image(file_obj)
    summary = await summarize_text(text)
    return summary

async def process_docx(file_obj):
    doc = docx.Document(file_obj)
    text = "\n".join([para.text for para in doc.paragraphs])
    summary = await summarize_text(text)
    return summary

async def process_photo(photo):
    file = await photo.get_file()
    file_content = await file.download_as_bytearray()
    text = await extract_text_from_image(BytesIO(file_content))
    summary = await summarize_text(text)
    return summary

async def process_audio(audio):
    file = await audio.get_file()
    file_content = await file.download_as_bytearray()
    audio_segment = AudioSegment.from_file(BytesIO(file_content))
    audio_segment.export("temp.wav", format="wav")
    
    recognizer = sr.Recognizer()
    with sr.AudioFile("temp.wav") as source:
        audio_data = recognizer.record(source)
        text = recognizer.recognize_google(audio_data)
    
    summary = await summarize_text(text)
    return summary

async def process_text(text):
    return text

async def process_url(url):
    response = requests.get(url)
    soup = BeautifulSoup(response.text, 'html.parser')
    text = soup.get_text()
    summary = await summarize_text(text)
    return summary

async def query_knowledge_base(query):
    # This function should be implemented to search your vector database
    # and return relevant information based on the query
    # For now, we'll just return a placeholder response
    return f"Here's what I found related to your query: '{query}'"