requirements.txt
python-telegram-bot
python-dotenv
openai
PyPDF2
python-docx
requests
beautifulsoup4
SpeechRecognition
pydub